+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs.Demo/Controllers/QueryController.cs
@@ -1,6 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Allegro.Extensions.Cqrs.Abstractions.Queries;
+using Allegro.Extensions.Cqrs.Demo.Queries;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Allegro.Extensions.Cqrs.Demo.Controllers;
@@ -16,23 +17,10 @@ public class QueryController : ControllerBase
         _queryDispatcher = queryDispatcher;
     }
 
-    [HttpGet]
-    public async Task<IActionResult> Execute([FromQuery] string id, CancellationToken cancellationToken)
+    [HttpGet("bar")]
+    public async Task<IActionResult> Bar([FromQuery] string? id, CancellationToken cancellationToken)
     {
-        var result = await _queryDispatcher.Query(new SampleQuery(id), cancellationToken);
+        var result = await _queryDispatcher.Query(new BarQuery(id), cancellationToken);
         return Ok(result);
-    }
-}
-
-internal record SampleQuery(string SomeId) : IQuery<SampleData>;
-
-public record SampleData(string SomeData);
-
-internal class SampleQueryHandler : IQueryHandler<SampleQuery, SampleData>
-{
-    public Task<SampleData?> Handle(SampleQuery query, CancellationToken cancellationToken)
-    {
-        // should take data directly from read model on dedicates sql query to view of some data
-        return Task.FromResult<SampleData?>(new SampleData("Some data 1"));
     }
 }