+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs/Commands/CommandDispatcher.cs
@@ -1,10 +1,12 @@
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 using Allegro.Extensions.Cqrs.Abstractions.Commands;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Allegro.Extensions.Cqrs.Commands;
 
+// TODO: can we extract common logic for command and query? Should we?
 internal sealed class CommandDispatcher : ICommandDispatcher
 {
     private readonly IServiceProvider _serviceProvider;
@@ -16,6 +18,11 @@ internal sealed class CommandDispatcher : ICommandDispatcher
     {
         // TODO: maybe some configuration to reuse outer scope instead of creating new one
         using var scope = _serviceProvider.CreateScope();
+
+        var commandValidators = scope.ServiceProvider.GetServices<ICommandValidator<TCommand>>();
+
+        await Task.WhenAll(commandValidators.Select(p => p.Validate(command)));
+
         var handler = scope.ServiceProvider.GetService<ICommandHandler<TCommand>>();
 
         if (handler is null)