+++ src/Allegro.Extensions.Cqrs/Allegro.Extensions.Cqrs/Queries/QueryDispatcher.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Linq;
+using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using Allegro.Extensions.Cqrs.Abstractions.Queries;
@@ -13,10 +15,36 @@ internal sealed class QueryDispatcher : IQueryDispatcher
     public QueryDispatcher(IServiceProvider serviceProvider)
         => _serviceProvider = serviceProvider;
 
-    public async Task<TResult?> Query<TResult>(IQuery<TResult> query, CancellationToken cancellationToken)
+    public async Task<TResult> Query<TResult>(IQuery<TResult> query, CancellationToken cancellationToken)
     {
         // TODO: maybe some configuration to reuse outer scope instead of creating new one
         using var scope = _serviceProvider.CreateScope();
+
+        var validatorType = typeof(IQueryValidator<>).MakeGenericType(query.GetType());
+        var queryValidators = scope.ServiceProvider.GetServices(validatorType);
+
+        var validateMethodInfo = validatorType
+            .GetMethod(nameof(IQueryValidator<IQuery<TResult>>.Validate));
+
+        if (validateMethodInfo is null)
+        {
+            throw new MissingMethodException($"Missing method Validate in validator for query {query.GetType().FullName}");
+        }
+
+        // https://learn.microsoft.com/en-us/dotnet/core/compatibility/core-libraries/7.0/reflection-invoke-exceptions
+
+        await Task.WhenAll(queryValidators.Select(p =>
+        {
+            // this is not a constructor, we can skip null
+            var invoke = validateMethodInfo.Invoke(
+                p,
+                BindingFlags.DoNotWrapExceptions,
+                null,
+                new object?[] { query, cancellationToken },
+                null)!;
+
+            return (Task)invoke;
+        }));
 
         // TODO: micro-optimization possibility - cache those types
         var handlerType = typeof(IQueryHandler<,>).MakeGenericType(query.GetType(), typeof(TResult));
@@ -31,14 +59,20 @@ internal sealed class QueryDispatcher : IQueryDispatcher
         var handleMethodInfo = handlerType
             .GetMethod(nameof(IQueryHandler<IQuery<TResult>, TResult>.Handle));
 
-        var invoke = handleMethodInfo?.Invoke(handler, new object?[] { query, cancellationToken });
-
-        if (invoke is null)
+        if (handleMethodInfo is null)
         {
-            return default;
+            throw new MissingMethodException($"Missing method Handle in handler for query {query.GetType().FullName}");
         }
 
-        return await (Task<TResult?>)invoke;
+        // this is not a constructor, we can skip null
+        var invoke = handleMethodInfo.Invoke(
+            handler,
+            BindingFlags.DoNotWrapExceptions,
+            null,
+            new object?[] { query, cancellationToken },
+            null)!;
+
+        return await (Task<TResult>)invoke;
     }
 }